src/requirements.txt
django==4.0.3
djangorestframework==3.13.1
requests==2.27.1

gunicorn==20.1.0

psycopg2-binary==2.9.3